This Fixes:
Issue 7

Please notice:
You only download the Objects of the MODERATOR not of any user 
conected to  whiteboard. So if there is not moderator any more
cause he has left the room you will not download the stuff of 
other people.
Also missing:
For audince modus something familiar has to be done but not implemented
yet cause it needs another logic (seperated Issue)

+ a Bug that prevents users from changing the Moderator