Fix problem with keyboard at login screen
In the login screen, the enter key is bind to "Sign in". It is problem when you
click on "not a member" or "Forgot password".
This patch fix this, "enter" is "sign in" "register" or "send mail with reset
link"
There is still a problem with error popup, as the key is still on the button
behind the popup. Popups are modal but in laszlo modal seem to be for mouse not
keyboard